requirements.txt
mkdocs
"mkdocstrings[python]"
mkdocs-material
pyyaml
mkdocs-toc-md